AdS5_3_1 [types]: add EvolPartition struct and some methods
after a lot of back and forth, i think this is the best way: use a structure to
hold an NTuple of *all* the arrays that are evolved in time, including the
boundary, gauge, and all the bulk ones across all u domains. this behaves as a
Vector a la ArrayPartition from RecursiveArrayTools.

now need to add some convenience methods to read from this to the structs that
are used to compute the bulk quantities