output: re-add remove_existing parameter
this deletes the output folder if present. it's a bit dangerous, since one can
completely remove data from a simulation. this was the reason we had removed
this in commit 4b56d9f649db7d7c4455b7cc3587966355b24888 in favour of the
overwrite parameter that was passed directly to the hdf5 routines.

still, it can be useful to have this strictly for fast debugging runs. so adding
it back and removing the overwrite parameter.