AdS5_3_1 [solve_nested]: bugfix! compute derivatives before calling the function set_innerBCs!
function set_innerBCs assumes that the derivatives of phi have been computed,
which was not the case! this commit fixes this.

this is only an actual problem when running with the scalar field, but even
without it there could be instances where the derivative of phi was randomly
initialized to NaNs, which would then propagate.