AdS5_3_1 [compute_gauge_t]: use mul! instead of mul_col! to build differential operators
the difference is not huge, but this way we can save a few copy operations and
we can do without the mul_col! operation, which we can then remove.

also remove comments about systems not being well-conditioned, since in fact
they are. we see this by calling cond(A_mat), where A_mat is the operator to be
inverted, instead of det(A_mat), which will always overflow for large matrices. see, eg, the discussion here: https://stackoverflow.com/questions/13145948/how-to-find-out-if-a-matrix-is-singular/