AdS5_3_1 [run]: add options to time integrator
in particular, the relative tolerance when using adaptive integrators is
important, as the default choice of 1e-3 was leading to very high timesteps in
the beginning of the integration -- which lead to some non-physical oscillations
in, eg, the energy as function of time at fixed x and y. so make a new parameter
called reltol that is 1e-6 by default, and which leads to much smoother results.

see here for description of options and algorithm:
https://diffeq.sciml.ai/latest/basics/common_solver_opts/
https://diffeq.sciml.ai/latest/extras/timestepping/

also, add a routine to estimate the initial timestep, so that we no longer need
to specify it (and can set dt = :auto). this routine gives only a very rough
estimate (it should be more of an upper limit, actually), but when using
adaptive integrators the algorithm should then be able to relax to a more
optimal choice.