operators for non-periodic finite differencing (#21)
* deriv_operators: use mod1 function instead of mod

much cleaner this way, since there is no need to shift vectors around

* deriv_operators: ensure that dx is a float and add check if needed to call mod1 function

* deriv_operators: add one more element to FiniteDiffDeriv and simplify some functions

* deriv_operators: use newly introduced element stencil_offset when taking the derivatives

this will more straightforwardly generalize to non-centered FD cases

* deriv_operators: split FD derivatives in routine doing only interior points and another doing only boundary points

* deriv_operators: remove method for cross-derivatives acting on Matrices

it's not needed since it's a particular case of the generic one acting on any
Array. it seems the same in terms of efficiency as well, so let's remove it as
it's superfluous.

* deriv_operators: split FD derivatives in two routines also for cross derivatives

* deriv_operators: add new abstract type, AbstractFiniteDiff, and new PeriodicFD struct

this should be useful to separate between periodic and non-periodic operators

* fornberg: add Fornberg routine to compute FD weights

ref: https://doi.org/10.1090/S0025-5718-1988-0935077-0
slighly adapted from DiffEqOperators module

* deriv_operators: use fornberg algorithm here, instead of hardcoded weights

* deriv_operators: use the methods explicitly with PeriodicFD when appropriate

* fornberg: use rational numbers to prevent loss of precision

* deriv_operators: add information about boundaries to FiniteDiffDeriv struct and add EqualSizeStencilFD struct to initialize it

TODO: add relevant methods

* deriv_operators: re-order high_boundary_offsets and fix method for FiniteDiffDeriv

* deriv_operators: fix getindex method for FiniteDiffDeriv

* deriv_operators [getindex]: use zero(T) instead of 0.0

* deriv_operators: start preparing latest methods for cross-derivatives

* test [deriv_operators_tests]: add some tests for EqualSizeStencilFD structs

* deriv_operators: remove mod1 function call in cases where it's not needed

* deriv_operators: add methods for cross derivatives of FiniteDiffDeriv

add also some tests

* WIP FD unit tests;

locally when Jecco is compiling it fails to see the
EqualSizeStencilFD. In repl "] st" shows that the Jecco
version at the end has "#AH_new". Does it mean that it is
the version from the AH_new branch?

Signed-off-by: Thanasis Giannakopoulos <giannakopoulos.th@gmail.com>

* FD unit tests;

fixed a sign prob.
for the Fornberg test it seems that if I do not explicitly include
that function in the tests, locally me compiled Jecco does not see it.

Signed-off-by: Thanasis Giannakopoulos <giannakopoulos.th@gmail.com>

* WIP FD unti tests:

the last deriv_operators_tests.jl fails, that is for crossed deriv.
with periodic and non-periodic operators. It seems that the problem is
for pointwise devs.

Signed-off-by: Thanasis Giannakopoulos <giannakopoulos.th@gmail.com>

* tests [deriv_operators_tests]: prefix calculate_weights call with "Jecco."

this is needed since the function is not exported

* deriv_operators: methods for crossed derivatives in interior point operations can be more general

ie, valid for periodic and non-periodic operators

* deriv_operators: add methods for cross derivatives with PeriodicFD and FiniteDiffDeriv types

* test [deriv_operators_tests]: update and add loops that cover the new methods

Co-authored-by: Thanasis Giannakopoulos <giannakopoulos.th@gmail.com>