output, input: fix errors and deprecation warnings coming from HDF5 library
some update in the HDF5 library had broken a few things for us.

now, we're explicitly adding the HDF5 package version used (0.14) to the
Project.toml file. we should have been doing this before, actually. so, to run
previous versions of the code, one would have to add, to the Project.toml file,
a previous version of the HDF5 library:

[compat]
HDF5  = "0.13"