filtering: add Kreiss-Oliger dissipation operators
so far, we've been applying the KO dissipation as convolution filters, but i
don't like it very much since the way we're doing is not functional at all, in
that the input gets overwritten.

this can have some non-intended consequences. during the time evolution, for
instance, i believe that in the Runge-Kutta substeps the functions are filtered
twice.

so let's try to add the dissipation in a more "traditional" way. and let's try
and mimic that in the exponential filtering...