AdS5_3_1/run: change logic when chosing integrator, to conform to changes in OrdinaryDiffEq
there used to be a good way to check if the integrator was defined in
OrdinaryDiffEq through isa(alg, OrdinaryDiffEq.OrdinaryDiffEqAlgorithm). but
this functionality was undocumented and at some point was removed. so now we
merely assume that, if it's not one of "our" algorithms, it must be an
OrdinaryDiffEq integrator. and if not, some error will appear.